requirements.txt
# NirnAI RAG Review Prototype - Dependencies
chromadb>=0.4.0
sentence-transformers>=2.2.0
pydantic>=2.0.0
python-dotenv>=1.0.0
openai>=1.0.0